requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
